rename tests to PhpDocInfo
