+++ packages/BetterPhpDocParser/tests/PhpDocInfo/PhpDocInfoSource/expected-with-replaced-type.txt
